Conversion of TaskReviewDAL to new Framework
* Move review metrics related methods from TaskReviewService/Repository to
  TaskReviewMetricsService/Repository

* Remove TaskReviewDAL completely and move all methods into TaskReviewService

* Fix SQL error in review search if user was not a super user and has no
  managed projects.
