+++ app/org/maproulette/framework/controller/TaskReviewController.scala
@@ -7,15 +7,19 @@ package org.maproulette.framework.controller
 import javax.inject.Inject
 import akka.util.ByteString
 import org.maproulette.data.ActionManager
+import org.maproulette.exception.NotFoundException
 import org.maproulette.framework.service.{
   ChallengeListingService,
   ChallengeService,
   ProjectService,
   TaskReviewService,
-  UserService
+  UserService,
+  ServiceManager
 }
 import org.maproulette.framework.psql.Paging
 import org.maproulette.framework.model.{Challenge, ChallengeListing, Project, User}
+import org.maproulette.framework.mixins.ParentMixin
+import org.maproulette.framework.repository.TaskRepository
 import org.maproulette.session.{SessionManager, SearchParameters, SearchTaskParameters}
 import org.maproulette.utils.Utils
 import play.api.mvc._
@@ -35,15 +39,231 @@ class TaskReviewController @Inject() (
     override val actionManager: ActionManager,
     override val bodyParsers: PlayBodyParsers,
     service: TaskReviewService,
+    taskRepository: TaskRepository,
     challengeListingService: ChallengeListingService,
     challengeService: ChallengeService,
     projectService: ProjectService,
     userService: UserService,
-    components: ControllerComponents
+    components: ControllerComponents,
+    serviceManager: ServiceManager
 ) extends AbstractController(components)
-    with MapRouletteController {
+    with MapRouletteController
+    with ParentMixin {
 
   implicit val challengeListingWrites: Writes[ChallengeListing] = Json.writes[ChallengeListing]
+
+  /**
+    * Gets and claims a task that needs to be reviewed.
+    *
+    * @param id Task id to work on
+    * @return
+    */
+  def startTaskReview(id: Long): Action[AnyContent] = Action.async { implicit request =>
+    this.sessionManager.authenticatedRequest { implicit user =>
+      val task = this.taskRepository.retrieve(id) match {
+        case Some(t) => t
+        case None    => throw new NotFoundException(s"Task with $id not found, cannot start review.")
+      }
+
+      val result = this.service.startTaskReview(user, task)
+      Ok(Json.toJson(result))
+    }
+  }
+
+  /**
+    * Releases a claim on a task that needs to be reviewed.
+    *
+    * @param id Task id to work on
+    * @return
+    */
+  def cancelTaskReview(id: Long): Action[AnyContent] = Action.async { implicit request =>
+    this.sessionManager.authenticatedRequest { implicit user =>
+      val task = this.taskRepository.retrieve(id) match {
+        case Some(t) => t
+        case None    => throw new NotFoundException(s"Task with $id not found, cannot cancel review.")
+      }
+
+      val result = this.service.cancelTaskReview(user, task)
+      Ok(Json.toJson(result))
+    }
+  }
+
+  /**
+    * Gets and claims the next task that needs to be reviewed.
+    *
+    * Valid search parameters include:
+    * cs => "my challenge name"
+    * o => "mapper's name"
+    * r => "reviewer's name"
+    *
+    * @return Task
+    */
+  def nextTaskReview(
+      onlySaved: Boolean = false,
+      sort: String,
+      order: String,
+      lastTaskId: Long = -1,
+      excludeOtherReviewers: Boolean = false
+  ): Action[AnyContent] = Action.async { implicit request =>
+    this.sessionManager.authenticatedRequest { implicit user =>
+      SearchParameters.withSearch { implicit params =>
+        // Cancel review claim of last task
+        try {
+          val lastTask = this.taskRepository.retrieve(lastTaskId) match {
+            case Some(t) => t
+            case None =>
+              throw new NotFoundException(s"Task with $lastTaskId not found, cannot cancel review.")
+          }
+          this.service.cancelTaskReview(user, lastTask)
+        } catch {
+          case _: Throwable => // do nothing, it's okay if the user didn't lock the prior task
+        }
+
+        val result = this.service.nextTaskReview(
+          user,
+          params,
+          onlySaved,
+          sort,
+          order,
+          (if (lastTaskId == -1) None else Some(lastTaskId)),
+          excludeOtherReviewers
+        )
+        val nextTask = result match {
+          case Some(task) =>
+            Ok(Json.toJson(this.service.startTaskReview(user, task)))
+          case None =>
+            throw new NotFoundException("No tasks found to review.")
+        }
+
+        nextTask
+      }
+    }
+  }
+
+  /**
+    * Gets tasks where a review is requested
+    *
+    * @param limit The number of tasks to return
+    * @param page The page number for the results
+    * @param sort The column to sort
+    * @param order The order direction to sort
+    * @param excludeOtherReviewers exclude tasks that have been reviewed by someone else
+    * @return
+    */
+  def getReviewRequestedTasks(
+      onlySaved: Boolean = false,
+      limit: Int,
+      page: Int,
+      sort: String,
+      order: String,
+      excludeOtherReviewers: Boolean = false,
+      includeTags: Boolean = false
+  ): Action[AnyContent] = Action.async { implicit request =>
+    this.sessionManager.userAwareRequest { implicit user =>
+      SearchParameters.withSearch { implicit params =>
+        //cs => "my challenge name"
+        //o => "mapper's name"
+        //r => "reviewer's name"
+        val (count, result) = this.service.getReviewRequestedTasks(
+          User.userOrMocked(user),
+          params,
+          onlySaved,
+          Paging(limit, page),
+          sort,
+          order,
+          true,
+          excludeOtherReviewers
+        )
+        Ok(
+          Json.obj(
+            "total" -> count,
+            "tasks" -> insertChallengeJSON(
+              this.serviceManager,
+              result,
+              includeTags
+            )
+          )
+        )
+      }
+    }
+  }
+
+  /**
+    * Gets reviewed tasks where the user has reviewed or requested review
+    *
+    * @param allowReviewNeeded Whether we should return tasks where status is review requested also
+    * @param limit The number of tasks to return
+    * @param page The page number for the results
+    * @param sort The column to sort
+    * @param order The order direction to sort
+    * @return
+    */
+  def getReviewedTasks(
+      allowReviewNeeded: Boolean = false,
+      limit: Int,
+      page: Int,
+      sort: String,
+      order: String,
+      includeTags: Boolean = false
+  ): Action[AnyContent] = Action.async { implicit request =>
+    this.sessionManager.userAwareRequest { implicit user =>
+      SearchParameters.withSearch { implicit params =>
+        val (count, result) = this.serviceManager.taskReview.getReviewedTasks(
+          User.userOrMocked(user),
+          params,
+          allowReviewNeeded,
+          limit,
+          page,
+          sort,
+          order
+        )
+        Ok(
+          Json.obj(
+            "total" -> count,
+            "tasks" -> insertChallengeJSON(
+              this.serviceManager,
+              result,
+              includeTags
+            )
+          )
+        )
+      }
+    }
+  }
+
+  /**
+    * Gets clusters of review tasks. Uses kmeans method in postgis.
+    *
+    * @param reviewTasksType Type of review tasks (1: To Be Reviewed 2: User's reviewed Tasks 3: All reviewed by users)
+    * @param numberOfPoints Number of clustered points you wish to have returned
+    * @param onlySaved include challenges that have been saved
+    * @param excludeOtherReviewers exclude tasks that have been reviewed by someone else
+    *
+    * @return A list of ClusteredPoint's that represent clusters of tasks
+    */
+  def getReviewTaskClusters(
+      reviewTasksType: Int,
+      numberOfPoints: Int,
+      onlySaved: Boolean = false,
+      excludeOtherReviewers: Boolean = false
+  ): Action[AnyContent] = Action.async { implicit request =>
+    this.sessionManager.userAwareRequest { implicit user =>
+      SearchParameters.withSearch { implicit params =>
+        Ok(
+          Json.toJson(
+            this.serviceManager.taskReview.getReviewTaskClusters(
+              User.userOrMocked(user),
+              reviewTasksType,
+              params,
+              numberOfPoints,
+              onlySaved,
+              excludeOtherReviewers
+            )
+          )
+        )
+      }
+    }
+  }
 
   /**
     * Returns a list of challenges that have reviews/review requests.
@@ -146,307 +366,6 @@ class TaskReviewController @Inject() (
             onlySaved
           )
         Ok(Json.toJson(results))
-      }
-    }
-  }
-
-  /**
-    * Gets reviewed tasks where the user has reviewed or requested review
-    *
-    * @param reviewTasksType - 1: To Be Reviewed 2: User's reviewed Tasks 3: All reviewed by users
-    * @param onlySaved Only include saved challenges
-    * @param excludeOtherReviewers exclude tasks that have been reviewed by someone else
-    * @return
-    */
-  def getReviewMetrics(
-      reviewTasksType: Int,
-      onlySaved: Boolean = false,
-      excludeOtherReviewers: Boolean = false,
-      includeByPriority: Boolean = false,
-      includeByTaskStatus: Boolean = false
-  ): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.userAwareRequest { implicit user =>
-      SearchParameters.withSearch { implicit params =>
-        val result = this.service.getReviewMetrics(
-          User.userOrMocked(user),
-          reviewTasksType,
-          params,
-          onlySaved,
-          excludeOtherReviewers
-        )
-
-        if (includeByPriority || includeByTaskStatus) {
-          var resultJson: JsValue = Json.obj(
-            "reviewActions" -> Json.toJson(result)
-          )
-
-          if (includeByPriority) {
-            val priorityMap = this.fetchPriorityReviewMetrics(
-              User.userOrMocked(user),
-              reviewTasksType,
-              params,
-              onlySaved,
-              excludeOtherReviewers
-            )
-
-            resultJson =
-              Utils.insertIntoJson(resultJson, "priorityReviewActions", Json.toJson(priorityMap))
-          }
-
-          if (includeByTaskStatus) {
-            val statusMap = this.fetchByTaskStatusReviewMetrics(
-              User.userOrMocked(user),
-              reviewTasksType,
-              params,
-              onlySaved,
-              excludeOtherReviewers
-            )
-
-            resultJson =
-              Utils.insertIntoJson(resultJson, "statusReviewActions", Json.toJson(statusMap))
-          }
-
-          Ok(resultJson)
-        } else {
-          Ok(Json.toJson(List(result)))
-        }
-      }
-    }
-  }
-
-  private def fetchPriorityReviewMetrics(
-      user: User,
-      reviewTasksType: Int,
-      params: SearchParameters,
-      onlySaved: Boolean,
-      excludeOtherReviewers: Boolean
-  ): scala.collection.mutable.Map[String, JsValue] = {
-    val prioritiesToFetch =
-      List(Challenge.PRIORITY_HIGH, Challenge.PRIORITY_MEDIUM, Challenge.PRIORITY_LOW)
-
-    val priorityMap = scala.collection.mutable.Map[String, JsValue]()
-
-    prioritiesToFetch.foreach(p => {
-      val newParams =
-        params.copy(taskParams = params.taskParams.copy(taskPriorities = Some(List(p))))
-
-      val pResult = this.service.getReviewMetrics(
-        user,
-        reviewTasksType,
-        newParams,
-        onlySaved,
-        excludeOtherReviewers
-      )
-
-      priorityMap.put(p.toString, Json.toJson(pResult))
-    })
-
-    priorityMap
-  }
-
-  private def fetchByTaskStatusReviewMetrics(
-      user: User,
-      reviewTasksType: Int,
-      params: SearchParameters,
-      onlySaved: Boolean,
-      excludeOtherReviewers: Boolean
-  ): scala.collection.mutable.Map[String, JsValue] = {
-    val statusesToFetch =
-      List(
-        Task.STATUS_FIXED,
-        Task.STATUS_FALSE_POSITIVE,
-        Task.STATUS_ALREADY_FIXED,
-        Task.STATUS_TOO_HARD
-      )
-
-    val statusMap = scala.collection.mutable.Map[String, JsValue]()
-
-    statusesToFetch.foreach(m => {
-      val newParams = params.copy(
-        taskParams = params.taskParams.copy(taskStatus = Some(List(m)))
-      )
-
-      val mResult = this.service.getReviewMetrics(
-        user,
-        reviewTasksType,
-        newParams,
-        onlySaved,
-        excludeOtherReviewers
-      )
-
-      statusMap.put(m.toString, Json.toJson(mResult))
-    })
-
-    statusMap
-  }
-
-  /**
-    * Returns a CSV export of review metrics per mapper.
-    *
-    * SearchParameters:
-    *   mappers Optional limit to reviews of tasks by specific mappers
-    *   reviewers Optional limit reviews done by specific reviewers
-    *   priorities Optional limit to only these priorities
-    *   startDate Optional start date to filter by reviewedAt date
-    *   endDate Optional end date to filter by reviewedAt date
-    * @param onlySaved Only include saved challenges
-    * @return
-    */
-  def extractMapperMetrics(
-      onlySaved: Boolean = false
-  ): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.userAwareRequest { implicit user =>
-      SearchParameters.withSearch { implicit params =>
-        val allReviewStatuses = List(
-          Task.REVIEW_STATUS_REQUESTED,
-          Task.REVIEW_STATUS_APPROVED,
-          Task.REVIEW_STATUS_REJECTED,
-          Task.REVIEW_STATUS_ASSISTED,
-          Task.REVIEW_STATUS_DISPUTED
-        )
-
-        val metrics = this.service.getMapperMetrics(
-          User.userOrMocked(user),
-          params.copy(
-            taskParams = SearchTaskParameters(
-              taskReviewStatus = Some(allReviewStatuses)
-            )
-          ),
-          onlySaved
-        )
-
-        val projectName =
-          params.getProjectIds match {
-            case Some(pId) =>
-              // Searching by project, just fetch name
-              pId
-                .map(
-                  this.projectService.retrieve(_) match {
-                    case Some(p) => p.displayName.get
-                    case None    => ""
-                  }
-                )
-                .mkString("|")
-            case None =>
-              params.getChallengeIds match {
-                case Some(cId) =>
-                  // We have a list of challenges. If these challenges all belong
-                  // to the same project we can use the parent project name.
-                  val challengeList = this.challengeService.list(cId)
-                  val parentProject: Option[Long] =
-                    if (challengeList.forall(_.general.parent == challengeList.head.general.parent))
-                      Some(challengeList.head.general.parent)
-                    else None
-                  parentProject match {
-                    case Some(pp) =>
-                      this.projectService.retrieve(pp) match {
-                        case Some(p) => p.displayName.get
-                        case None    => ""
-                      }
-                    case None => ""
-                  }
-                case None => ""
-              }
-          }
-
-        val challengeName =
-          params.getChallengeIds match {
-            case Some(cIds) =>
-              this.challengeService.list(cIds).map(_.name).mkString("|")
-            case None => ""
-          }
-
-        val mapperNames =
-          this.userService
-            .retrieveListById(metrics.map(m => m.userId.get), Paging())
-            .map(u => u.id -> u.name)
-            .toMap
-
-        val byReviewStatusMetrics =
-          allReviewStatuses
-            .map(reviewStatus => {
-              val reviewStatusMetrics = this.service.getMapperMetrics(
-                User.userOrMocked(user),
-                params.copy(
-                  taskParams = SearchTaskParameters(
-                    taskReviewStatus = Some(List(reviewStatus))
-                  )
-                ),
-                onlySaved
-              )
-
-              reviewStatus -> (reviewStatusMetrics.map(m => m.userId.get -> m).toMap)
-            })
-            .toMap
-
-        val seqString = metrics.map(row => {
-          var mapper = mapperNames.get(row.userId.get)
-
-          val result = new StringBuilder(
-            s"${mapper.get},${projectName},${if (challengeName == "") "" else s"${challengeName},"}" +
-              s",${row.total},,,${row.reviewRequested}," +
-              s"${row.reviewApproved},${row.reviewRejected},${row.reviewAssisted}," +
-              s"${row.reviewDisputed},${row.fixed},${row.falsePositive},${row.alreadyFixed}," +
-              s"${row.tooHard}"
-          )
-
-          allReviewStatuses.foreach(rs => {
-            if (byReviewStatusMetrics.get(rs).get.contains(row.userId.get)) {
-              val rsRow         = byReviewStatusMetrics.get(rs).get(row.userId.get)
-              val rsTimeSeconds = Math.round(rsRow.avgReviewTime / 1000)
-              val rsPercent     = Math.round(rsRow.total * 100 / row.total)
-              result ++=
-                s"\n${mapper.get},${projectName},${if (challengeName == "") ""
-                else s"${challengeName},"}" +
-                  s"${Task.reviewStatusMap.get(rs).get},${rsRow.total}," +
-                  s"${rsPercent},${rsTimeSeconds},,,,," +
-                  s",${rsRow.fixed},${rsRow.falsePositive},${rsRow.alreadyFixed}," +
-                  s"${rsRow.tooHard}"
-            }
-          })
-          result.toString
-        })
-
-        Result(
-          header = ResponseHeader(
-            OK,
-            Map(CONTENT_DISPOSITION -> s"attachment; filename=mapper_review_metrics.csv")
-          ),
-          body = HttpEntity.Strict(
-            ByteString(
-              s"Mapper,Project,${if (challengeName == "") "" else "Challenge,"}" +
-                s"Review Status,Total Review Tasks,Coverage %,Avg Review Time (seconds)," +
-                s"Review Requested,Approved,Needs Revision,Approved w/Fixes,Contested," +
-                s"${Task.STATUS_FIXED_NAME},${Task.STATUS_FALSE_POSITIVE_NAME}," +
-                s"${Task.STATUS_ALREADY_FIXED_NAME},${Task.STATUS_TOO_HARD_NAME}\n"
-            ).concat(ByteString(seqString.mkString("\n"))),
-            Some("text/csv; header=present")
-          )
-        )
-      }
-    }
-  }
-
-  /**
-    * Returns a breakdown of tag metrics
-    *
-    * @return
-    */
-  def getReviewTagMetrics(
-      reviewTasksType: Int,
-      onlySaved: Boolean,
-      excludeOtherReviewers: Boolean
-  ): Action[AnyContent] = Action.async { implicit request =>
-    this.sessionManager.userAwareRequest { implicit user =>
-      SearchParameters.withSearch { implicit params =>
-        val result = this.service.getReviewTagMetrics(
-          User.userOrMocked(user),
-          reviewTasksType,
-          params,
-          onlySaved,
-          excludeOtherReviewers
-        )
-        Ok(Json.toJson(result))
       }
     }
   }
