+++ app/org/maproulette/framework/psql/Order.scala
@@ -8,7 +8,12 @@ package org.maproulette.framework.psql
 import anorm.NamedParameter
 import org.maproulette.framework.psql.filter.SQLClause
 
-case class OrderField(name: String, direction: Int = Order.DESC, table: Option[String] = None)
+case class OrderField(
+    name: String,
+    direction: Int = Order.DESC,
+    table: Option[String] = None,
+    isColumn: Boolean = true
+)
 
 /**
   * Basic class to handle any ordering that needs to be added to the query
@@ -20,7 +25,9 @@ case class Order(fields: List[OrderField] = List.empty) extends SQLClause {
     val directions = fields.map(field => field.direction).distinct
     val orderFields = fields
       .map(field => {
-        SQLUtils.testColumnName(field.name)
+        if (field.isColumn) {
+          SQLUtils.testColumnName(field.name)
+        }
         val fieldDirection = if (directions.size > 1) Order.direction(field.direction) else ""
         val columnKey = field.table.getOrElse(table) match {
           case ""    => field.name
