+++ app/org/maproulette/framework/repository/TaskRepository.scala
@@ -8,21 +8,76 @@ package org.maproulette.framework.repository
 import java.sql.Connection
 import scala.concurrent.duration.FiniteDuration
 
-import anorm.SqlParser.{get, scalar}
+import anorm.SqlParser.{get, scalar, str}
 import anorm.ToParameterValue
 import anorm.{RowParser, ~}
 import anorm._, postgresql._
 import javax.inject.{Inject, Singleton}
+import org.maproulette.Config
+import org.maproulette.framework.mixins.TaskParserMixin
 import org.maproulette.framework.psql.Query
 import org.maproulette.framework.psql.filter.{BaseParameter, CustomParameter}
 import org.maproulette.models.Task
-import org.maproulette.framework.model.User
+import org.maproulette.framework.model.{User, Project}
+import org.maproulette.cache.CacheManager
 import play.api.db.Database
 import play.api.libs.json._
 
 @Singleton
-class TaskRepository @Inject() (override val db: Database) extends RepositoryMixin {
+class TaskRepository @Inject() (override val db: Database, config: Config)
+    extends RepositoryMixin
+    with TaskParserMixin {
   implicit val baseTable: String = "tasks"
+
+  // The cache manager for tasks
+  val cacheManager =
+    new CacheManager[Long, Task](config, Config.CACHE_ID_TASKS)(taskReads, taskReads)
+
+  /**
+    * For a given id returns the task
+    *
+    * @param id The id of the task you are looking for
+    * @param c An implicit connection, defaults to none and one will be created automatically
+    * @return None if not found, otherwise the Task
+    */
+  def retrieve(id: Long): Option[Task] = {
+    this.cacheManager.withCaching { () =>
+      this.withMRConnection { implicit c =>
+        val query = s"SELECT $retrieveColumnsWithReview FROM ${this.baseTable} " +
+          "LEFT OUTER JOIN task_review ON task_review.task_id = tasks.id " +
+          "WHERE tasks.id = {id}"
+        SQL(query)
+          .on(Symbol("id") -> id)
+          .as(this.getTaskParser(this.updateAndRetrieve).singleOpt)
+      }
+    }(id)
+  }
+
+  /**
+    * Allows us to lazy update the geojson data
+    *
+    * @param taskId The identifier of the task
+    */
+  def updateAndRetrieve(
+      taskId: Long,
+      geojson: Option[String],
+      location: Option[String],
+      cooperativeWork: Option[String]
+  ): (String, Option[String], Option[String]) = {
+    geojson match {
+      case Some(g) => (g, location, cooperativeWork)
+      case None =>
+        this.withMRTransaction { implicit c =>
+          SQL("SELECT * FROM update_geometry({id})")
+            .on(Symbol("id") -> taskId)
+            .as((str("geo") ~ get[Option[String]]("loc") ~ get[Option[String]]("fix_geo")).*)
+            .headOption match {
+            case Some(values) => (values._1._1, values._1._2, values._2)
+            case None         => throw new Exception("Failed to retrieve task data")
+          }
+        }
+    }
+  }
 
   /**
     * Updates the completionResponses on a Task
