+++ conf/v2_route/review.api
@@ -14,7 +14,7 @@
 #     in: path
 #     description: The id of the task to fetch and claim
 ###
-GET     /task/:id/review/start                       @org.maproulette.controllers.api.TaskReviewController.startTaskReview(id:Long)
+GET     /task/:id/review/start                       @org.maproulette.framework.controller.TaskReviewController.startTaskReview(id:Long)
 ###
 # tags: [ Review ]
 # summary: Cancels a claim on a task for review
@@ -31,7 +31,7 @@ GET     /task/:id/review/start                       @org.maproulette.controller
 #     in: path
 #     description: The id of the task to cancel claim on
 ###
-GET     /task/:id/review/cancel                       @org.maproulette.controllers.api.TaskReviewController.cancelTaskReview(id:Long)
+GET     /task/:id/review/cancel                       @org.maproulette.framework.controller.TaskReviewController.cancelTaskReview(id:Long)
 ###
 # tags: [ Review ]
 # summary: Retrieves tasks that need review
@@ -77,7 +77,7 @@ GET     /task/:id/review/cancel                       @org.maproulette.controlle
 #     in: query
 #     description: The search string used to match the name of the person requesting the review. (review_requested_by)
 ###
-GET     /tasks/review                          @org.maproulette.controllers.api.TaskReviewController.getReviewRequestedTasks(onlySaved: Boolean ?= false, limit:Int ?= -1, page:Int ?= 0, sort:String ?= "", order:String ?= "ASC", excludeOtherReviewers: Boolean ?= false, includeTags: Boolean ?= false)
+GET     /tasks/review                          @org.maproulette.framework.controller.TaskReviewController.getReviewRequestedTasks(onlySaved: Boolean ?= false, limit:Int ?= -1, page:Int ?= 0, sort:String ?= "", order:String ?= "ASC", excludeOtherReviewers: Boolean ?= false, includeTags: Boolean ?= false)
 ###
 # tags: [ Review ]
 # summary: Retrieves reviewed tasks that have been reviewed either by this user or where the user requested
@@ -130,7 +130,7 @@ GET     /tasks/review                          @org.maproulette.controllers.api.
 #     in: query
 #     description: In response include list of tags for each task
 ###
-GET     /tasks/reviewed                          @org.maproulette.controllers.api.TaskReviewController.getReviewedTasks(allowReviewNeeded:Boolean ?= false, limit:Int ?= 10, page:Int ?= 0, sort:String ?= "", order:String ?= "ASC", includeTags:Boolean ?= false)
+GET     /tasks/reviewed                          @org.maproulette.framework.controller.TaskReviewController.getReviewedTasks(allowReviewNeeded:Boolean ?= false, limit:Int ?= 10, page:Int ?= 0, sort:String ?= "", order:String ?= "ASC", includeTags:Boolean ?= false)
 ###
 # tags: [ Review ]
 # summary: Retrieves and claims a the next review needed Task
@@ -168,7 +168,7 @@ GET     /tasks/reviewed                          @org.maproulette.controllers.ap
 #     in: query
 #     description: The search string used to match the Reviewer names. (reviewed_by)
 ###
-GET     /tasks/review/next                       @org.maproulette.controllers.api.TaskReviewController.nextTaskReview(onlySaved:Boolean ?= false, sort:String ?= "", order:String ?= "ASC", lastTaskId:Long ?= -1, excludeOtherReviewers:Boolean ?= false)
+GET     /tasks/review/next                       @org.maproulette.framework.controller.TaskReviewController.nextTaskReview(onlySaved:Boolean ?= false, sort:String ?= "", order:String ?= "ASC", lastTaskId:Long ?= -1, excludeOtherReviewers:Boolean ?= false)
 ###
 # tags: [ Review ]
 # summary: Retrieves nearby Tasks
@@ -242,7 +242,7 @@ GET     /tasks/review/nearby/:proximityId           @org.maproulette.framework.c
 #     in: query
 #     description: Also include a breakdown of review status by task status
 ###
-GET     /tasks/review/metrics                          @org.maproulette.framework.controller.TaskReviewController.getReviewMetrics(reviewTasksType: Int, onlySaved:Boolean ?= false, excludeOtherReviewers: Boolean ?= false, includeByPriority:Boolean ?= false, includeByTaskStatus:Boolean ?= false)
+GET     /tasks/review/metrics                          @org.maproulette.framework.controller.TaskReviewMetricsController.getReviewMetrics(reviewTasksType: Int, onlySaved:Boolean ?= false, excludeOtherReviewers: Boolean ?= false, includeByPriority:Boolean ?= false, includeByTaskStatus:Boolean ?= false)
 ###
 # tags: [ Review ]
 # summary: Retrieves task review clusters
@@ -276,7 +276,7 @@ GET     /tasks/review/metrics                          @org.maproulette.framewor
 #     in: query
 #     description: exclude tasks that have been reviewed by someone else
 ###
-GET     /taskCluster/review                                @org.maproulette.controllers.api.TaskReviewController.getReviewTaskClusters(reviewTasksType:Int, points:Int ?= 100, onlySaved:Boolean ?= false, excludeOtherReviewers:Boolean ?= false)
+GET     /taskCluster/review                                @org.maproulette.framework.controller.TaskReviewController.getReviewTaskClusters(reviewTasksType:Int, points:Int ?= 100, onlySaved:Boolean ?= false, excludeOtherReviewers:Boolean ?= false)
 ###
 # tags: [ Review ]
 # summary: Retrieve a summary of review coverage for mappers
@@ -304,7 +304,7 @@ GET     /taskCluster/review                                @org.maproulette.cont
 #     in: query
 #     description: Only show challenges that have been saved.
 ###
-GET     /tasks/review/mappers/export                          @org.maproulette.framework.controller.TaskReviewController.extractMapperMetrics(onlySaved:Boolean ?= false)
+GET     /tasks/review/mappers/export                          @org.maproulette.framework.controller.TaskReviewMetricsController.extractMapperMetrics(onlySaved:Boolean ?= false)
 ###
 # tags: [ Review ]
 # summary: Retrieves metrics grouped by tag
@@ -329,4 +329,4 @@ GET     /tasks/review/mappers/export                          @org.maproulette.f
 #     in: query
 #     description: exclude tasks that have been reviewed by someone else
 ###
-GET     /tasks/review/tag/metrics                          @org.maproulette.framework.controller.TaskReviewController.getReviewTagMetrics(reviewTasksType: Int, onlySaved:Boolean ?= false, excludeOtherReviewers: Boolean ?= false)
+GET     /tasks/review/tag/metrics                          @org.maproulette.framework.controller.TaskReviewMetricsController.getReviewTagMetrics(reviewTasksType: Int, onlySaved:Boolean ?= false, excludeOtherReviewers: Boolean ?= false)
