+++ test/org/maproulette/framework/FrameworkMasterSuite.scala
@@ -36,8 +36,11 @@ class FrameworkMasterSuite extends Suites with BeforeAndAfterAll with TestDataba
     new TagServiceSpec,
     new TagRepositorySpec,
     new TaskRepositorySpec,
+    new TaskClusterRepositorySpec,
+    new TaskReviewRepositorySpec,
     new TaskServiceSpec,
     new TaskReviewServiceSpec,
+    new TaskReviewMetricsServiceSpec,
     new UserMetricsServiceSpec,
     new UserSavedObjectsServiceSpec,
     new UserSavedObjectsRepositorySpec,
